requirements.txt
Django
django-debug-toolbar
djangorestframework
psycopg2